% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRTvsRB_FPKM}
\alias{getRTvsRB_FPKM}
\title{Calculate the rate of transcription and translation.}
\usage{
getRTvsRB_FPKM(stage, totalCounts, boundCounts, unboundCounts, totalFPKM,
  cluster, cutoff = 1)
}
\arguments{
\item{stage}{Developmental stage; \code{string}.}

\item{totalCounts}{Read counts from total RNA sequencing experiment.}

\item{boundCounts}{Read counts form polysome profiling bound fraction
sequencing.}

\item{unboundCounts}{Read counts form polysome profiling unbound fraction
sequencing.}

\item{totalFPKM}{FPKM values from total RNA sequencing experiment.}

\item{cluster}{Clustering information for RNA fractions.}

\item{cutoff}{FPKM cutoff value.}
}
\value{
A data frame with rates and clustering information
}
\description{
Integrates the count from multiple sequencing experiments to produce the
rates of transcription and translation in contrast to the egg developmental
stage.
}
\seealso{
Materials and Methods, Winata CL, Development, 11 December 2017
\url{http://dev.biologists.org/content/early/2017/12/08/dev.159566}
}
