% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotLogMat}
\alias{plotLogMat}
\title{Plot the scatterplot}
\usage{
plotLogMat(data, xlim = c(-5, 5), ylim = c(-5, 5), title = NULL,
  type = "histogram", bins = 100, colour = c("#ef8a62", "#67a9cf"))
}
\arguments{
\item{data}{A data frame for plotting.}

\item{xlim, ylim}{X and Y axis limits.}

\item{title}{Plot title.}

\item{type}{Type of marginal density plot passed to ggMarginal function.}

\item{bins}{Number of histogram bins.}

\item{colour}{Colour values for different clusters.}
}
\value{
A ggplot2 plot object.
}
\description{
Uses ggplot2 and ggExtra to produce the scatterplot with density plots on the
margins.
}
