% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare the data for plotting}
\usage{
prepare_data(data, xlim, ylim, title)
}
\arguments{
\item{data}{A data frame for operations}

\item{xlim, ylim}{X and Y axis limits.}

\item{title}{Dataset name}
}
\value{
A processed data frame
}
\description{
Filter maternal genes and subset the rows to match the limits of the graph
}
\keyword{internal}
